pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "arch-analyser"
version = "0.1.0"
description = "5-prompt Plan-and-Execute architecture analyser"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
arch-analyser = "arch_analyser.cli:main"